EFCore1.0.0重构以及新驱动适配

+++ src/Microsoft.EntityFrameworkCore.Specification.Tests/TestModels/Inheritance/Kiwi.cs
@@ -8,7 +8,7 @@ namespace Microsoft.EntityFrameworkCore.Specification.Tests.TestModels.Inheritan
         public Island FoundOn { get; set; }
     }
 
-    public enum Island : int
+    public enum Island : byte
     {
         North,
         South
